Add support for Netlify to contact form
This adds support for using Netlify's form service
and enables it by default. It also takes advantage of their spam
protection. Lastly, it facilitates setting a name for your contact form
in case you have more than one as the form name is utilized by Netlify.

The Netlify parts of this are based on their docs at:

- https://docs.netlify.com/forms/setup/
- https://docs.netlify.com/forms/spam-filters/