Add parameter to validate ownership in Mastodon
When utilized, this will produce the rel=me link that Mastodon looks for
when verifying a user owns a link listed in their profile.

The site verification partial is also updated so that comments are only
rendered for sections that are actually used.

A social icon for Mastodon is also added into the example site's config